added ability to edit folders

+++ app/app/api/folder_routes.py
@@ -1,8 +1,9 @@
 from flask import Blueprint, request
 from flask_login import login_required, current_user
+import datetime
 
 from app.models import db, Folder
-from app.forms import NewFolderForm
+from app.forms import NewFolderForm, EditFolderForm
 
 folder_routes = Blueprint('folders', __name__)
 
@@ -21,7 +22,7 @@ def form_errors(validation_errors):
 def create_folder():
     form = NewFolderForm()
     form["csrf_token"].data = request.cookies["csrf_token"]
-    print(form)
+
     if form.validate_on_submit():
         folder = Folder(
             name=form.data['name'],
@@ -34,12 +35,28 @@ def create_folder():
     return {'errors': form_errors(form.errors)}
 
 
-@folder_routes.route('/<int:id>', methods=['DELETE'])
+@folder_routes.route('/<int:id>', methods=['DELETE', 'PUT', 'GET'])
 @login_required
-def delete_folder(id):
+def update_folder(id):
     folder = Folder.query.get(id)
-    db.session.delete(folder)
-    db.session.commit()
+
+    if request.method == 'DELETE':
+        db.session.delete(folder)
+        db.session.commit()
+
+    elif request.method == 'PUT':
+        form = EditFolderForm()
+        form["csrf_token"].data = request.cookies["csrf_token"]
+
+        if form.validate_on_submit():
+           folder.name = form.data["name"]
+           folder.category_id = form.data["category"]
+           folder.updated_at = datetime.datetime.utcnow()
+           db.session.commit()
+
+    elif request.method == 'GET':
+        return folder.to_dict()
+
 
     user_folders = Folder.query.filter_by(user_id=current_user.id)
